+++ sdk/src/main/java/com/iovation/launchkey/sdk/transport/domain/ServerSentEventAuthorizationResponseCore.java
@@ -16,6 +16,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class ServerSentEventAuthorizationResponseCore {
     private final String auth;
     private final String userPushId;
@@ -24,7 +25,6 @@ public class ServerSentEventAuthorizationResponseCore {
     private final String publicKeyId;
 
     @JsonCreator
-    @JsonIgnoreProperties(ignoreUnknown = true)
     public ServerSentEventAuthorizationResponseCore(
             @JsonProperty(value = "auth", required = true) String auth,
             @JsonProperty(value = "user_push_id", required = true) String userPushId,